requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-WTF
email-validator